feat: validate files for duplicates TDE-711 (#404)
If duplicates are found, the process will stop with an error message.
Detailed information about the duplicates is written to a file which
will be easy to find in the argo ui/workflow artifacts bucket rather
than printed in the logs.

[topo-workflows draft
PR](https://github.com/linz/topo-workflows/pull/127) with screenshots of
test runs.

A [new ticket](https://toitutewhenua.atlassian.net/browse/TDE-744) has
been created to rename the files at this step if no duplicates are
found. Then we can remove the duplicated code in topo-imagery.

---------

Co-authored-by: Blayne Chard <bchard@linz.govt.nz>