feat(group): increase logging for grouping (#609)
#### Why?

When artifact errors occur in argo its hard to figure out exactly what
files were written and what is in those files.

#### Description

This increases the logging to explain exactly which files were written
and some information about their contents.

When running argo containers having the exact github action that made
the container is very useful, we were loosing the -attempt number so it
was not possible to get the exact version number.