fix(copy): copy and fix content-types TDE-859 (#633)
A lot of files we write into S3 use the default content-type of
"application/octet-stream" this content-type is not very useful for the
browser to know to render the file or download it.

A lot of these files are written by other systems and we cannot correct
in all systems.

This adds 

```
--fix-content-type=true
--fix-content-type=false // default
```

Which will attempt to guess what the content type should be when copying
files,