feat: pretty print JSON files TDE-759 (#647)
### Description
Add a `pretty-print` command that format all JSON files within a
directory using `prettier` with a copy of the `linzjs` prettier
configuration. A target directory can be specified instead of
overwritting the existing files. This is only for testing purpose as it
does not handle duplicate filenames.

### Intention
Some of the STAC files (`Item`) are not pretty printed.

### Checklist
If not applicable, provide explanation of why.

- [x] Tests updated
- [x] Docs updated
- [x] Issue linked in Title