feat: lint imagery input target paths TDE-857 (#730)
#### Motivation

in order to be sure that workflow input public taraget paths follow
conventions by linting workflow inputs where possible.

#### Modification

Adding lint-paths step to argo-tasks which will be used in `pubish-odr`
and can also be added to `publish-copy`

Currently just the publish paths are linted (linz-imagery & nz-imagery)
- this was based on the ticket description but more can be added if
required.
The dataset section of the s3 path is currently not linted as this is
too variable to check?

(this change will be used in a topo-workflow PR currently WIP)

#### Checklist

_If not applicable, provide explanation of why._

- [x] Tests updated
- [x] Docs updated
- [x] Issue linked in Title