feat(basemaps): Get vector title from the collection.json file. BM-1017 (#970)
#### Motivation
We are populating title inside collection.json file from the etl. The
`collection.title` is the combined title when processing multiple vector
layers which is using the parameter from `etl --title`. And all the
`collection.links` should contain the `lds:title` from the lds export.
We should using this title when there is only one lds layer exists in
the etl output.

#### Modification
Get the title from collection.json and get the title from `lds:title`
for individual vector output.

#### Checklist

- [ ] Tests updated - no test
- [ ] Docs updated - no doc
- [x] Issue linked in Title