feat: verify multihash when copy TDE-1181 TDE-1172 (#1001)
#### Motivation

When copying files across S3 bucket, we need a way to verify if the file
already exists in order to prevent overriding it. The current way to
verify that is based on the file size which is not accurate.

#### Modification

Add a verification on the file checksum, stored in its metadata or
generated on the go.

#### Checklist

_If not applicable, provide explanation of why._

- [x] Tests updated
- [ ] Docs updated (N/A)
- [x] Issue linked in Title