feat: generate path from slug TDE-1319 (#1136)
#### Motivation

The previous logic to construct the ODR path has partially been replaced
by the `linz:slug` field that is constructed in the [stac-setup Argo
Tasks
command](https://github.com/linz/argo-tasks/tree/master/src/commands/stac-setup).
`generate-path` can be simplified to use the pre-constructed slug when
publishing to the ODR.

#### Modification

- Read the `linz:slug` field from the standardised collection.json file
and use this to construct the ODR path.
- Remove unused functions from the `generate-path` command.
- Refactor where it makes sense to reduce duplication of code.

#### Checklist

- [x] Tests updated
- [x] Docs updated
- [x] Issue linked in Title

---------

Co-authored-by: paulfouquet <86932794+paulfouquet@users.noreply.github.com>
Co-authored-by: Blayne Chard <bchard@linz.govt.nz>