fix: remove GSD unit from stac setup command inputs TDE-1339 TDE-1211 (#1153)
#### Motivation

A change was made recently to accept the GSD in the format without a
trailing `m` e.g. `0.3` rather than `0.3m`. However, the old format
should still be accepted as this may accidentally be entered as a
workflow parameter.

#### Modification

Small refactor to match refactor made to `generate-path` command when
handling `GeospatialDataCategories`.
Add a function `formatGsd` to check the format of the GSD, remove a
trailing `m` and log a warning.

#### Checklist

- [x] Tests updated
- [x] Docs updated
- [x] Issue linked in Title

---------

Co-authored-by: Blayne Chard <bchard@linz.govt.nz>