feat: use `derived_from` checksum to see if tile needs re-creating TDE-1449 (#1207)
#### Motivation

As a data maintainer, when updating a national DEM Hillshade, I would
like to recreate the hillshade tiles only for the nationally combined
DEM tiles that have changed from the previous update.

#### Modification

Added command `identifyUpdatedItems` to identify for which items in a
collection.json file there are updated items in the respective source
collections (via derived_from links).

#### Checklist

_If not applicable, provide explanation of why._

- [x] Tests updated
- [x] Docs updated
- [x] Issue linked in Title