feat: compress and delete source files TDE-1482 (#1216)
**Note:** feedback especially sought on:
- using cpu count to decide concurrency
- optimisation/refactoring of code to reduce any code duplication and
improve efficiency

#### Motivation

The requirement to move data to a designated archive bucket with a
Glacier Deep Archive lifecycle policy in order to retain data not
frequently accessed for the medium/long term at a lower cost. The source
data would be deleted and noncurrent versions expired after a set date.

#### Modification

- Give copy workflow the ability to compress destination data
- Give copy workflow the ability to delete source data

#### Checklist

- [x] Tests updated
- [x] Docs updated
- [x] Issue linked in Title

---------

Co-authored-by: Alice Fage <afage@linz.govt.nz>