fix: identify-updated-items should return ordered input based on collection order TDE-1588 (#1241)
### Motivation

The order of the `input` was random (see multiple automated test runs
failing for both `should only add modified items to file-list.json` and
`should add all source items if no target collection is provided`). It
needs to follow the order of the collection passed in
`--source-collections`.

### Modifications

- Explicitly sort the `input` arrays to match the order of
`--source-collections` as provided by the user, ensuring that files from
the first collection always appear first, then the second, etc.

### Verification

Run the automated tests locally a bunch of times.