feat: tileindex-validate to output GSD TDE-1655 (#1268)
### Motivation

As an imagery maintainer
I want to make sure the resolution of all TIFF images matches
so that the data is consistent and accurate (in line with the metadata)

This change is to validate all TIFFs have a consistent GSD and make the
GSD available to other workflow steps.

### Modifications

Extract GSD (resolution) from all tiff files that get processed.
Output GSD to a file for use by argo workflows

Added tests.

### Verification

Unit tests.