feat(stac-collection-output): get scale from ODR dataset TDE-1673 (#1282)
### Motivation

The new `tileindex-validate` command behaviour requires the scale of
existing datasets to override automatic tiling, if necessary.

### Modifications

Added a new command `stac-collection-output` to read the ODR collection
and output the scale so it can be used as an Argo Workflows parameters.
Other parameters can be added later if necessary.

### Verification

Unit tests and workflow run with PR container.