feat(tileindex-validate)!: automatically select scale of output tiles TDE-1674 (#1287)
### Motivation

Input TIFFs should be automatically tiled using their GSD and imagery
type to determe the appropriate output scale. It avoids data managers to
know what the supplied tile size is, what should be the output tile
size, and whether or not a "retiling" (output tile made of several input
TIFFs) is needed.
Validation of perfect alignment and duplicates does not make sense at
this stage as the tiling will always occurs when `--scale=auto`. This
validation feature has been removed and will be part of a visual QA
feature next.

### Modifications

- :warning: **BREAKING CHANGE** `--retile` and `--validate` flags has
been removed
- Add an `auto` mode to the `--scale` flag
- Always validate inputs have consistent `GSD`
<!-- TODO: Say what changes you made. -->

<!-- TODO: Attach screenshots if you changed the UI. -->

### Verification

Automated and manual tests
<!-- TODO: Say how you tested your changes. -->