fix: get GSD from TFW sidecar files TDE-1715 (#1292)
### Motivation

Users no longer supply the GSD value as a workflow parameter, it is
sourced from the TIFFs. However, some TIFFs provided are not GeoTIFFs
and do not contain the resolution (GSD). They are supplied with a
sidecar TFW file that contains this information. We need to read the
resolution from the TFW sidecar file in the same way as is done for the
bbox if necessary.

### Modifications

Added a function to get the resolution from the TFW file and some minor
refactoring to reduce code duplication and allow the use of an async
function.

### Verification

Unit tests, local testing, and a workflow run against the PR container.